+++ src/Atc.Cosmos.EventStore/Cosmos/GuidEventIdProvider.cs
@@ -1,9 +1,0 @@
-using Atc.Cosmos.EventStore.Events;
-
-namespace Atc.Cosmos.EventStore.Cosmos;
-
-internal class GuidEventIdProvider : IEventIdProvider
-{
-    public string CreateUniqueId(IStreamMetadata metadata)
-        => Guid.NewGuid().ToString();
-}