+++ src/Atc.Cosmos.EventStore/Events/EventBatchProducer.cs
@@ -5,16 +5,13 @@ namespace Atc.Cosmos.EventStore.Events;
 internal class EventBatchProducer : IEventBatchProducer
 {
     private readonly IDateTimeProvider dateTimeProvider;
-    private readonly IEventIdProvider eventIdProvider;
     private readonly IEventNameProvider nameProvider;
 
     public EventBatchProducer(
         IDateTimeProvider dateTimeProvider,
-        IEventIdProvider eventIdProvider,
         IEventNameProvider nameProvider)
     {
         this.dateTimeProvider = dateTimeProvider;
-        this.eventIdProvider = eventIdProvider;
         this.nameProvider = nameProvider;
     }
 
@@ -58,20 +55,18 @@ internal class EventBatchProducer : IEventBatchProducer
         string? causationId,
         DateTimeOffset timestamp)
     {
-        var eventId = eventIdProvider.CreateUniqueId(metadata);
         var streamId = metadata.StreamId.Value;
         var name = nameProvider.GetName(evt);
 
         return new EventDocument<object>
         {
-            Id = eventId,
+            Id = $"{version}",
             PartitionKey = streamId,
             Data = evt,
             Properties = new EventMetadata
             {
                 CausationId = causationId,
                 CorrelationId = correlationId,
-                EventId = eventId,
                 StreamId = streamId,
                 Version = version,
                 Timestamp = timestamp,