+++ src/Atc.Cosmos.EventStore/Events/IEventIdProvider.cs
@@ -1,6 +1,0 @@
-namespace Atc.Cosmos.EventStore.Events;
-
-public interface IEventIdProvider
-{
-    string CreateUniqueId(IStreamMetadata metadata);
-}