+++ test/Atc.Cosmos.EventStore.Cqrs.Tests/Internal/ConsumeAnyEventAsyncMetadataTests.cs
@@ -45,7 +45,6 @@ public class ConsumeAnyEventAsyncMetadataTests :
                 .Should()
                 .BeEquivalentTo(
                     new EventMetadata(
-                        fixture.ConsumableEven.Metadata.EventId,
                         EventStreamId.FromStreamId(fixture.ConsumableEven.Metadata.StreamId),
                         fixture.ConsumableEven.Metadata.Timestamp,
                         (long)fixture.ConsumableEven.Metadata.Version,