+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added custom event data converters to be configured using `EventStoreOptions`. This will enable scenarioes such as converting from one version of an event to another.
 - Unknown or invalid events can now be observed through the `IConsumeEvent<T>` and `IConsumeEventAsync<T>` but using well known types `FaultedEvent` and `UnknownEvent`.
 - Introduced new interfaces `IConsumeAnyEvent` and `IConsumeAnyEventAsync` for consuming any event without specifying it type.
+- Fixed raise condition when 2 command processors tries to add the first event to the same stream concurrently.
 
 ### Removed
 - Setting `ConfigurationString` when configuring event store options.
+- `EventId` has been removed from `Metadata`.
 
 ## [1.6.8] - 2022-07-06